requirements.txt
Flask==2.3.3
aiohttp==3.8.6
asyncio==3.4.3
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
requests==2.31.0
openai==0.28.1
schedule==1.2.0
python-wordpress-xmlrpc==2.3
textstat==0.7.3
yake==0.4.8
python-crontab==2.7.1